backend/requirements.txt
langchain
langchain-cohere
cohere
faiss-cpu
python-dotenv
Flask
flask-cors

